The read-side fix landed independently on master in #11094, which rewrote
LightTxDecoder onto RlpReader/RlpWriter and already decodes the proof version
with DecodeByte. Kept master's decoder verbatim and carried over only the
on-disk byte pin, which master's tests still lack.

+++ src/Nethermind/Nethermind.Db/LogIndex/LogIndexConfig.cs
@@ -13,7 +13,7 @@ public class LogIndexConfig : ILogIndexConfig
     public bool Reset { get; set; } = false;
 
     // set from PruningConfig via decorator
-    public int? MaxReorgDepth { get; set; }
+    public ulong? MaxReorgDepth { get; set; }
 
     public int MaxBatchSize { get; set; } = 256;
     public int MaxAggregationQueueSize { get; set; } = 16;
